% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ao_cache_dir}
\alias{ao_cache_dir}
\title{Local cache directory for downloaded ONS lookup tables}
\usage{
ao_cache_dir()
}
\value{
A file path, created if it doesn't already exist.
}
\description{
Local cache directory for downloaded ONS lookup tables
}
\keyword{internal}
