% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ao_data_sources}
\alias{ao_data_sources}
\title{The data sources this package can fetch}
\usage{
ao_data_sources()
}
\value{
A character vector of source names, for use as \code{source} in
\code{\link[=ao_data]{ao_data()}}.
}
\description{
The data sources this package can fetch
}
\examples{
ao_data_sources()
}
