% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreasOfOpportunity-package.R
\docType{package}
\name{AreasOfOpportunity-package}
\alias{AreasOfOpportunity}
\alias{AreasOfOpportunity-package}
\title{AreasOfOpportunity: Filter and Map UK Geographic Areas}
\description{
A foundation for working with UK geographic areas. Lists and filters areas at Output Area, LSOA, Ward, Local Authority District, County, Community Safety Partnership and Police Force Area level (via the ONS Open Geography Portal), and fetches area boundaries for mapping. Intended as a base layer for other area-level data (e.g. crime, deprivation) to be joined on top of.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulWHeselwood/AreasOfOpportunity}
  \item Report bugs at \url{https://github.com/PaulWHeselwood/AreasOfOpportunity/issues}
}

}
\author{
\strong{Maintainer}: Paul Heselwood \email{paul.heselwood@proton.me}

Authors:
\itemize{
  \item Paul Heselwood \email{paul.heselwood@proton.me}
}

}
\keyword{internal}
