% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{check_level}
\alias{check_level}
\title{Validate a level argument}
\usage{
check_level(level)
}
\arguments{
\item{level}{A single string.}
}
\value{
\code{level}, invisibly, if valid; errors otherwise.
}
\description{
Validate a level argument
}
\keyword{internal}
