% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{sql_escape}
\alias{sql_escape}
\title{Escape a value for use in an ArcGIS REST \code{where} clause}
\usage{
sql_escape(x)
}
\arguments{
\item{x}{A single string.}
}
\value{
\code{x} with single quotes doubled, ready to wrap in \code{'...'}.
}
\description{
Escape a value for use in an ArcGIS REST \code{where} clause
}
\keyword{internal}
