% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{arcgis_query_sf}
\alias{arcgis_query_sf}
\title{Fetch matching features from an ArcGIS \code{FeatureServer} layer as \code{sf}}
\usage{
arcgis_query_sf(service, where, out_fields = "*")
}
\arguments{
\item{service}{Service name (relative to \link{ons_host}).}

\item{where}{SQL \code{where} clause.}

\item{out_fields}{Character vector of fields to return.}
}
\value{
An \code{sf} object with one row per matching feature.
}
\description{
Fetch matching features from an ArcGIS \code{FeatureServer} layer as \code{sf}
}
\keyword{internal}
